Squashed 'software/src/master/' changes from 8e7b695..6534d12
6534d12 Merge 'software/src/8.0' development into 'software/src/8.1' code base.
d470ed0 Delete 'bin' sub-directory from software/src/8.1.
131063e Merge branch 'master' (solaris-11.3) into macport
22eb2bd Merge branch 'master' of github.com:MichaelJCaruso/vision
10a5e73 Null merge of branch 'solaris-11.3-x86' into macport
88b4a5b Merge 8.1 '-U_FORTIFY_SOURCE' 'master' changes into 'solaris-11.3-x86'
ba3c43e Merge 8.1 '-U_FORTIFY_SOURCE' 'master' changes into 'macport'
eaa0212 Update 8.1 to use -U_FORTIFY_SOURCE Linux compiler option.
c462cfb .gitignore extra files generated by 'c++ -g' on the Mac.
c7c3ee7 Merge branch '8.1-bootstrap-fix' into staging
7220a0c Restore definition of 'setColumnPtr:' debug method.
5309049 Duplicate ad5c6309e763f9112e3c138beeb4b3fb101b6f4c changes under 8.1.
49d5efd Fix at-exit crashes of 8.1 executables on linux.
78ac3a0 Add output of 'visionBuilder debug' to .gitignore.
a9d6a01 Vision 8.1 for github!!

git-subtree-dir: software/src/master
git-subtree-split: 6534d124b1fab97bd52ce4cf87a27434bbb4b6ff

+++ src/kernel/V_VAllocator.h
@@ -146,7 +146,7 @@ namespace V {
     //  Allocation
     private:
 	static size_t allocationSize (size_t sCell) {
-	    unsigned int nCells = 1024 / sCell;
+	    unsigned int nCells = static_cast<unsigned int>(1024 / sCell);
 	    return nCells > 0 ? nCells * sCell : sCell;
 	}
 	static pointer_t getSpace (size_t sCell);
@@ -319,7 +319,7 @@ namespace V {
     //  Grain and Counter Indices
     protected:
 	static unsigned int cellIndex (size_t sObject) {
-	    return sObject <= CellFloor ? 0 : (sObject - CellFloor + CellGrain - 1) / CellGrain;
+	    return sObject <= CellFloor ? 0 : static_cast<unsigned int>((sObject - CellFloor + CellGrain - 1) / CellGrain);
 	}
 	static size_t cellSize (unsigned int xGrain) {
 	    return CellGrain * (xGrain + 1) + CellFloor;
